llama-index-integrations/storage/chat_store/llama-index-storage-chat-store-yugabytedb/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "docker>=7.1.0,<8",
    "ipython==8.10.0",
    "jupyter>=1.0.0,<2",
    "mypy==0.991",
    "pre-commit==3.2.0",
    "pylint==2.15.10",
    "pytest==8.3.0",
    "pytest-asyncio==0.21.0",
    "pytest-mock==3.11.1",
    "ruff==0.11.11",
    "types-Deprecated>=0.1.0",
    "types-PyYAML>=6.0.12.12,<7",
    "types-protobuf>=4.24.0.4,<5",
    "types-redis==4.5.5.0",
    "types-requests==2.28.11.8",
    "types-setuptools==67.1.0.0",
    "black[jupyter]<=23.9.1,>=23.7.0",
    "codespell[toml]>=v2.2.6",
    "diff-cover>=9.2.0",
    "pytest-cov>=6.1.1",
]

[project]
name = "llama-index-storage-chat-store-yugabytedb"
version = "0.3.0.1"
description = "llama-index storage-chat-store yugabytedb integration"
authors = [{name = "YugabyteDB", email = "pypi@yugabyte.com"}]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = "MIT"
dependencies = [
    "llama-index-core>=0.13.0,<0.14",
    "psycopg2-yugabytedb>=2.9.3.5",
    "sqlalchemy-yugabytedb>=1.0.0.1",
]

[tool.codespell]
check-filenames = true
check-hidden = true
skip = "*.csv,*.html,*.json,*.jsonl,*.pdf,*.txt,*.ipynb"

[tool.hatch.build.targets.sdist]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.hatch.build.targets.wheel]
include = ["llama_index/"]
exclude = ["**/BUILD"]

[tool.llamahub]
contains_example = false
import_path = "llama_index.storage.chat_store.yugabytedb"

[tool.llamahub.class_authors]
YugabyteDBChatStore = "llama-index"

[tool.mypy]
disallow_untyped_defs = true
exclude = ["_static", "build", "examples", "notebooks", "venv"]
ignore_missing_imports = true
python_version = "3.8"
